% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteContinuousMI.R
\name{calculateDiscreteContinuousMI_Discretise}
\alias{calculateDiscreteContinuousMI_Discretise}
\title{calculate mutual information between a categorical value (X) and a continuous value (Y) using a discretisation and infotheo}
\usage{
calculateDiscreteContinuousMI_Discretise(
  df,
  groupXVar,
  valueYVar,
  bins = nrow(df)^(1/3)
)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{groupXVar}{- the column of the categorical value (X)}

\item{valueYVar}{- the column of the continuous value (Y)}

\item{bins}{- the number of bins (see infotheo::discretize)}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group a mutual information column (I). If df was not grouped this will be a single entry
}
\description{
calculate mutual information between a categorical value (X) and a continuous value (Y) using a discretisation and infotheo
}
