% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionFunctions.R
\name{ConditionalDistribution}
\alias{ConditionalDistribution}
\title{Get a named class distribution wrapper}
\description{
Get a named class distribution wrapper

Get a named class distribution wrapper
}
\keyword{distributions}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{names}}{the class name}

\item{\code{probabilities}}{the marginal probability of the class}

\item{\code{df}}{the distribution functions}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{names}}{the class name}

\item{\code{probabilities}}{the marginal probability of the class}

\item{\code{df}}{the distribution functions}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-withDistribution}{\code{ConditionalDistribution$withDistribution()}}
\item \href{#method-withRandomDistributions}{\code{ConditionalDistribution$withRandomDistributions()}}
\item \href{#method-sample}{\code{ConditionalDistribution$sample()}}
\item \href{#method-getPdf}{\code{ConditionalDistribution$getPdf()}}
\item \href{#method-plot}{\code{ConditionalDistribution$plot()}}
\item \href{#method-numericalMI}{\code{ConditionalDistribution$numericalMI()}}
\item \href{#method-theoreticalMI}{\code{ConditionalDistribution$theoreticalMI()}}
\item \href{#method-theoreticalMean}{\code{ConditionalDistribution$theoreticalMean()}}
\item \href{#method-theoreticalVariance}{\code{ConditionalDistribution$theoreticalVariance()}}
\item \href{#method-clone}{\code{ConditionalDistribution$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withDistribution"></a>}}
\subsection{Method \code{withDistribution()}}{
adds in a Distribution with a class name and weight
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionalDistribution$withDistribution(
  weight,
  distribution,
  class = distribution$label()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weight}}{the relative weight of this class}

\item{\code{distribution}}{the pdf as an R6 Distribution object (Distribution$new(fn, fnParams...))}

\item{\code{class}}{the classname}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withRandomDistributions"></a>}}
\subsection{Method \code{withRandomDistributions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionalDistribution$withRandomDistributions(n = 2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\subsection{Method \code{sample()}}{
produce a set of samples conforming to this distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionalDistribution$sample(n = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{the number of samples}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data frame of samples (labelled x) associated with classes (labelled "class")
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPdf"></a>}}
\subsection{Method \code{getPdf()}}{
plot this dictributions as pdf and cdf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionalDistribution$getPdf(xmin, xmax, resolution = 1001)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a datafram containing all classes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
plot this dictributions as pdf and cdf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionalDistribution$plot(xmin, xmax)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a ggassemble plot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numericalMI"></a>}}
\subsection{Method \code{numericalMI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionalDistribution$numericalMI(xmin, xmax, resolution = 1001)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-theoreticalMI"></a>}}
\subsection{Method \code{theoreticalMI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionalDistribution$theoreticalMI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-theoreticalMean"></a>}}
\subsection{Method \code{theoreticalMean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionalDistribution$theoreticalMean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-theoreticalVariance"></a>}}
\subsection{Method \code{theoreticalVariance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionalDistribution$theoreticalVariance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionalDistribution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
